# src/main/resources/openapi.yaml
openapi: "3.0.2"
info:
  title: Example Loans API
  version: "1.0"
servers:
  - url: localhost:8080
paths:
  /schedules/{id}:
    get:
      operationId: getSchedule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleResponse'
        default:
          description: error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorResponse'

  /schedules:
    get:
      operationId: getSchedulesList
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleListResponse'
        default:
          description: error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorResponse'
    post:
      operationId: addNewSchedule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SchedulePostBody"
      responses:
        '200':
          description: OK
        default:
          description: error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorResponse'
components:
  schemas:
    SchedulePostBody:
      type: object
      properties:
        assetValue:
          type: number
          format: double
        deposit:
          type: number
          format: double
        annualInterestPcnt:
          type: number
          format: double
        loanDurationMonths:
          type: integer
        balloonAmount:
          type: number
          format: double
    ScheduleResponse:
      type: object
      allOf:
        - $ref: "#/components/schemas/ScheduleListResponseItem"
        - type: array
          items:
            $ref: "#/components/schemas/ScheduleItem"


    ScheduleListResponseItem:
      type: object
      properties:
        id:
          type: integer
        scheduleRequestDetails:
          $ref: "#/components/schemas/SchedulePostBody"
        monthlyRepayment:
          type: number
          format: double
        totalInterestDue:
          type: number
          format: double
        totalPaymentsDue:
          type: number
          format: double

    ScheduleItem:
      type: object
      properties:
        period:
          type: integer
        payment:
          type: number
          format: double
        priciple:
          type: number
          format: double
        interest:
          type: number
          format: double
        balance:
          type: number
          format: double

    ScheduleListResponse:
      type: object
      properties:
        Schedules:
          type: array
          items:
            $ref: "#/components/schemas/ScheduleListResponseItem"

    errorResponse:
      type: object
      properties:
        message:
          type: string